fix(l2): fix zkVM programs (#1901)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

zkVM programs fail to compile after #1861 was merged to main. This PR
fixes the problem and bumps provers' versions.

L2 breaking changes keep being merged because the prover compilation job
isn't run in pushes that don't modify the L2 code, but the L2 is
strongly dependent on L1 code. This PR enables the job for all pushes.
The job takes ~13 minutes to complete, which is comparable to other jobs
which run on all pushes.

<!-- A clear and concise general description of the changes this PR
introduces -->

- fixes `ethrex-levm` adding c-kzg into the zkVMs programs dependency
tree
- bumps SP1 and Risc0 versions
- enables prover CI job for all pushes