ci(levm): add pipefail option to stop job execution if compilation fails (#1924)
**Motivation**

Currently, when the LEVM's CI compiles the project, the following
command is run:

```
make run-evm-ef-tests-ci | tee test_result.txt
```
However, if the compilation fails, the job is _not_ stopped. 

This is because, by default, bash will return the last exit code in the
pipeline. In this case, that is the `tee` command, which exits
correctly.

**Description**
Setting the `pipefail` option, will make it so that job will fail if the
`make` routine fails.


Closes #1918