fix(l1): discarding already connected peers (#1904)
**Motivation**
When running `make localnet`, we noticed the following error: 
```shell
Hello messages exchange failed: (Disconnect requested: Already connected)
```

**Description**
The problem is that the node starts a connection with us as receivers,
and we also try to start one as initiators. Whichever happens first will
provoke the second one and cause a disconnection because the other node
will detect it as already connected. This pr is a quick fix that simply
matches the `error` type, and if it detects `Already Connected` then it
does not discard the peer from the table (as there is still a `tcp` conn
open). A more proper fix would be to introduce a mutex when starting
connections.

Closes None