fix(l1): peers disconnecting after some time (#1923)
**Motivation**
After some time, when reaching about 10-15 peers we noticed all of them
would disconnect abruptly.

**Description**
The problem was that we were assuming disconnections when the stream
buffer was of length 0. While this may indicate an `EOF` thus
disconnection, that assumption turns out to be wrong as stated in the
[tokio
docs](https://docs.rs/tokio-util/latest/tokio_util/codec/struct.Framed.html#implementations):

> Note that, for some byte sources, the stream can be resumed after an
EOF by reading from it, even after it has returned `None`. Repeated
attempts to do so, without new data available, continue to return `None`
without creating more (closing) frames.


When the handshake hasn't been established we assume a disconnection but
after the handshake was established, then we don't assume a
disconnection and instead let the `check_periodic_task` detect
disconnection when sending the ping message.

What was interesting and we were not able to debug is why this would
happen when being connected to 10-15 peers :man_shrugging:.

Closes None