feat(l1): parallel state sync + trie rebuild in background + early storage healing (#1936)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
Currently, snap sync is working fine on small networks and local test
environments but it is still too slow to catch up to a
larger network such as holesky and sepolia. This PR aims to speed up the
state sync phase by dividing the account state and spawning multiple
processes to fetch multiple sections of the state at the same time.
While working on this PR I also noticed that the trie rebuild time is
quite significant so this PR also includes rebuilding the state &
storage tries in a permanent background process while the state sync is
active.
The PR also includes starting storage healing earlier for those storages
that are left in the fetcher's queue when the pivot becomes stale. This
change should alleviate the workload of the storage healer in the
post-state sync healing phase.

**Description**

***State Sync Paralellization***
- Create Snapshot DB tables to store fetched accounts and storages
during state sync
- Write accounts to the snapshot instead of the tries during state sync
- Divide the account state into N segments and fetch all segments at the
same time (The number of segments is currently set to 2 but can be
adjusted by changing a constant in the storage module)

***Trie Rebuild in Background***
- Add a background state trie rebuild process that will continuously try
to rebuild the state trie by reading the latest accounts from the state
snapshot, finishing when state sync is finished and there are no more
accounts in the state snapshot
- Add a background storage trie rebuild process that will listen for
incoming messages from the storage fetcher, which will advertise
storages that have been fully fetched, and rebuild the full storage
tries. It will end when state sync has finished and no more storages are
left in queue.
- Pass a `CancellationToken` to the background rebuild processes so that
they will not be stuck when taking down the node via Ctr+C

***Early Healing***
- The storage healer will be spawned before state sync as opposed to
after.
- A sender to the storage healer's channel will be passed to the storage
fetcher so that it can send unfetched storages after a stale pivot is
detected.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number