fix(levm): increase support for `stPreCompiledContracts` and `stPreCompiledContracts2` (#1969)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

In this PR we check if the fork is enabled to execute precompiles and
implement the gas cost for the `call` opcode in
[EIP-161](https://eips.ethereum.org/EIPS/eip-161)

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

* In the Fork Byzantium the precompiles `modexp`, `ecadd`, `ecmul` and
`ecpairing` are introduced. In the previous forks this precompiles must
be disable.
* For the `call` opcode in the previous forks to Spurious Dragon, the
empty account cost only checks for the account to be empty. On Spurious
Dragon and after, the value to send must be non zero to be charged with
the gas cost.

<!-- Link to issues: Resolves #111, Resolves #222 -->

The current state is:
* `stPreCompiledContracts: 298/299 (99.67%)`
* `stPreCompiledContracts2: 648/762 (85.04%)`

---------

Co-authored-by: Tomas Fabrizio Orsi <tomas.orsi@lambdaclass.com>
Co-authored-by: Federico Borello <156438142+fborello-lambda@users.noreply.github.com>