refactor(l1): modularize sync code & add doc comments for each module (#1961)
**Motivation**
The code for sync has grown significantly since its start, and having it
all under the same module can make it very hard to follow.
This PR aims to solve this by separating each different job into
different modules and adding module-level documentation for each,
explaining how each part of the sync works and integrates to the system
as a whole
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Separate sync code into different modules, one for each separate
functionality
* Add doc comments for each individual module explaining its purpose and
mechanics
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Notes for merging:
This PR will likely conflict with #1936 but it has the same exact code,
only moving it across modules, so it should be safe to solve all
conflicts by keeping the changes from this PR alone. Alternatively, you
can also merge this PR into #1936 as it has no conflicts and then merge
#1936

---------

Co-authored-by: Martin Paulucci <martin.c.paulucci@gmail.com>