feat(levm): add a warning when REVM fails an EF-test and ignore tests (#1985)
**Motivation**

There is no clear warning when REVM fails an EF-Test, one has to check
manually that there is a difference between REVM's post-state-root and
the expected.

**Description**

This PR adds a warning in the report when REVM fails a test. This should
also make it easier to check/parse how many tests REVM fails.

Also ignore some EF-tests that test wrong/impossible behavior.