fix(l1): hive invalid ancestor (#1926)
**Motivation**
Fixes hive invalid ancestor tests under `engine-cancun` with the pattern
`engine-cancun/Invalid Missing .* .*CanonicalReOrg=False`.

**Description**
The following was implemented to fix the tests:
- Because `forkchoice_update` and `new_payload` require us to send the
latest valid hash, we now cache the last valid hash of the chain
corresponding to bad blocks, these are the `invalid_ancestors`.
- The bad blocks are cached when:
   1. When failing to add a block during syncing
   2.  When the block execution of `new_payload` fails.
- The latest valid hash was fixed on `new_payload`; this was a known
issue see #982.
- Finally, when syncing, if we have a pending block from a `new_payload`
request, we attach it if the parent hash of the block corresponds to the
last header parent hash.

Closes #982
Advances on #1285