fix(l1): update `rpc-compat` hive tests to Prague and fix failing ones (#2043)
**Description**

This PR updates the hive tests suite to a new hive revision and fixes
new tests by removing the `total_difficulty` field on the block scheme.
This field was already removed in #304 and reintroduced in #566 due to
the usage in hive engine test suite.

Hive tests related to the wrong encoding of `nonce` field were also
fixed

---------

Co-authored-by: Julian Ventura <julian.ventura@lambdaclass.com>