fix(l2): add test chain with a non-empty block and bug fixes (#2068)
**Motivation**

The current `l2-loadtest.rlp` had empty blocks. After replacing with a
chain that contains a non-empty block, a bug was discovered when
creating a `ExecutionDB` from a `Store`: we would try to filter out new
accounts using revm's `AccountStatus` but for some reason every account
has the same status (`Touched`), meaning that we were not filtering at
all.

The solution was to not filter and assume that the `Store` is correct
based on the success of the pre-execution (so if an account is missing
from the parent state, then that account was created in the block to
prove).

**Description**

- replaces `l2-loadtest.rlp` file with one that contains a non empty
block
- fixes the bug described above