refactor(l1): make Blockchain an Arc<Blockchain> (#2110)
**Motivation**

This is a follow up to #2089 
Blockchain struct is supposed to be a singleton, we shouldn't be cloning
it, only passing references.

**Description**

- Remove `derive(Clone)` from Blockchain. Make arguments to
functions/structs `Arc<Blockchain>`, the only initialization of
blockchain is done in `ethrex.rs` inside an Arc.
- Remove `derive(Clone)` from mempool as the mempool shouldn't be cloned
outside of Blockchain, remove Arc from Mempool members as they should
always be inside the blockchain which already is Arc.
- Remove `pub` from storage inside of blockchain as it should only be
used by functions of `Blockchain`.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #2105