feat(l2): change config from `.env` to `config.toml` (#2100)
**Motivation**

Currently, users have to write the `.env` file by hand. The file is in a
format that is hard to read and will only get worse if more variables
get added to it.

**Description**

This PR makes it so that users now have to edit a `config.toml` instead.
That file gets converted into the `.env` file which is used by the L2.

It also provides a `config_example.toml` file, just like `.env.example`
before it.

Furthermore, this PR also changes where the files are mouunted in
Docker. Before they were mounted in root the Docker image's root `/`;
now the various files and directories are saved in the PWD (ATTOW,
`/usr/bin/local/`). Regarding the PWD, that has been moved to an
Makefile variable, that allows both the Dockerfile and the
docker-compose file to share the same variable definition. **There is no
default to that variable**, this is intentional to avoid hard to debug
situations where one file has one definition, while the other does not.

Sidenote: It also changes the `Genesis` `alloc` field to a `BTreeMap` so
it's deterministic in between runs.

Closes #1989 
Closes #2070

---------

Signed-off-by: Tomas Fabrizio Orsi <tomas.orsi@lambdaclass.com>
Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>
Co-authored-by: Estéfano Bargas <estefano.bargas@fing.edu.uy>
Co-authored-by: JereSalo <jeresalo17@gmail.com>
Co-authored-by: Francisco Krause Arnim <56402156+fkrause98@users.noreply.github.com>