chore(l1): create a `L1_CLIENT` environment variable for hive tests. (#2175)
**Motivation**

Currently, if you want to run the hive test suite with a different
client different than ethrex, you have to change the Makefile manually.

**Description**

This PR introduces a new environment variable `L1_CLIENT` that allows
you to select a different L1 to run the Hive tests with; without needing
to modify the Makefile manually.

---------

Signed-off-by: Tomas Fabrizio Orsi <tomas.orsi@lambdaclass.com>