test(levm): ensure behavior pairing check (#2160)
**Motivation**

This PR is for ensuring that the precompile bls_pairing_check works as
expected.
If any of the points in the pairs is a infinity point, that specific
pair should be skipped.

**Description**

We add a unit test that test this behavior. If a pair with a infinity
point is added, the result must not change.