l2(feat): new L2 proposer (#2167)
**Motivation**

Currently the L2 proposer uses the EngineAPI for creating blocks and
advancing the blockchain via three calls: `forkChoiceUpdate`,
`getPayload` and `newPayload`. This means that transactions get executed
twice (once when retrieving txs from the mempool and then when
validating the payload).

Also other components benefit from the execution result of the proposer,
like the comitter and prover, so this PR adds an execution result cache
that the proposer will update for every new block and other components
can consume it. The comitter uses this cache (for calculating state
diffs) and the prover will also in a future PR. This avoids re-execution
in these components.

closes #2114