feat(l2): make deposits for the rich accounts (#2095)
**Motivation**

In this PR we change the behavior of the deployer. We make deposits from
the rich accounts to the L1 in the L2 common bridge to make the rich
accounts of the L2.

**Description**

* We add a `make_deposits()` function in the `deployer.rs`. Here we
iterate from the Private Keys of the L1, calculate their addresses and
make a deposit
* We add a new file `private_keys_l1`
* We add new volumes and environment variables to the
`docker-compose-l2.yaml` for the files of the private keys and genesis
to be able to read them
* We remove the extra accounts from the `genesis-l2.json`

Closes #2017

---------

Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>
Co-authored-by: fborello-lambda <federicoborello@lambdaclass.com>
Co-authored-by: Federico Borello <156438142+fborello-lambda@users.noreply.github.com>